requirements.txt
flask==3.1.2
python-dotenv==1.2.1
sqlalchemy==2.0.46
requests==2.32.4
gunicorn==24.1.1
fpdf2==2.8.5
cryptography==46.0.3
beautifulsoup4==4.14.3
lxml==6.0.2
flask-login==0.6.3
werkzeug==3.1.5
flask-mail==0.10.0
flask-limiter==4.1.1
redis
pytest
pytest-cov
pytest-mock
flask-apscheduler==1.13.1
psycopg2-binary==2.9.11
alembic==1.18.4
prometheus-flask-exporter
flasgger==0.9.7.1
flask-cors==6.0.2
stripe==14.2.0
reportlab>=4.0
itsdangerous>=2.1